Refact cells and create UIImageView extension
