+++ challangeMarvel/challangeMarvel/CharacterDetails/CharacterDetailsInteractor.swift
@@ -10,7 +10,6 @@ import Foundation
 
 protocol CharacterDetailsInteractorProtocol {
     func onViewLoad()
-    func getCharacterImage(url: String)
 }
 
 class CharacterDetailsInteractor: CharacterDetailsInteractorProtocol {
@@ -31,14 +30,6 @@ class CharacterDetailsInteractor: CharacterDetailsInteractorProtocol {
         guard let collectionURI = character.comics?.collectionURI else { return }
         worker.getComics(url: collectionURI) { comics in
             self.presenter.show(comics: comics)
-        }
-    }
-    
-    func getCharacterImage(url: String) {
-        worker.getCharacterImage(url: url) { image in
-            if let image = image {
-                self.presenter.show(image: image)
-            }
         }
     }
 }
