+++ challangeMarvel/challangeMarvel/CharacterDetails/CharacterDetailsWorker.swift
@@ -10,7 +10,6 @@ import Foundation
 import UIKit
 
 protocol CharacterDetailsWorkerProtocol: class {
-    func getCharacterImage(url: String, completion: @escaping ((UIImage?) -> Void))
     func getComics(url: String, completion: @escaping (([Comic]) -> Void))
     func decode(data: Data) -> [Comic]?
 }
@@ -19,21 +18,6 @@ class CharacterDetailsWorker: CharacterDetailsWorkerProtocol {
     
     let httpClient = HTTPClient()
     let authentication = Authentication()
-    
-    func getCharacterImage(url: String, completion: @escaping ((UIImage?) -> Void)) {
-        if let apiUrl = URL(string: url) {
-            httpClient.get(url: apiUrl) { result -> Void in
-                switch result {
-                case .failure:
-                    break
-                case .success(let data):
-                    let loadedImage = UIImage(data: data)
-                    completion(loadedImage)
-                    break
-                }
-            }
-        }
-    }
     
     func getComics(url: String, completion: @escaping (([Comic]) -> Void)) {
         let urlCredentials = authentication.addCredentials(url: url)
